refactor: 컴포넌트 폴더 구조 계층 상세화
