+++ src/app.js
@@ -2,11 +2,11 @@ import "./global.css";
 import Filter from "./components/Filter";
 import Header from "./components/Header";
 import Tab from "./components/Tab";
-import ModalContainer from "./components/modal/ModalContainer";
-import RestaurantList from "./components/restaurants/RestaurantList";
+import Modal from "./components/Modal";
+import RestaurantList from "./components/RestaurantList";
 import { $ } from "./util/dom";
 
-const modal = new ModalContainer($(".modal"));
+const modal = new Modal($(".modal"));
 
 const header = new Header($(".gnb"));
 const tab = new Tab($(".restaurant-tab-container"));
