+++ src/components/RestaurantList/index.js
@@ -1,8 +1,8 @@
 import "./index.css";
-import RestaurantListManager from "../../../domain/RestaurantListManager";
-import { $ } from "../../../util/dom";
-import LocalStorage from "../../../util/LocalStorage";
-import RestaurantItem from "../RestaurantItem";
+import RestaurantListManager from "../../domain/RestaurantListManager";
+import { $ } from "../../util/dom";
+import LocalStorage from "../../util/LocalStorage";
+import RestaurantItem from "./RestaurantItem";
 
 class RestaurantList {
   $target;
